Support for validating sigs from ECDSA in tests
