+++ Phantasma.Pay/WalletUtils.cs
@@ -1,6 +1,9 @@
 ﻿using System;
+using System.IO;
 using Phantasma.Cryptography;
 using Phantasma.Pay.Chains;
+using Phantasma.Core.Utils;
+using Phantasma.Numerics;
 
 namespace Phantasma.Pay
 {
@@ -53,5 +56,20 @@ namespace Phantasma.Pay
                 default:  throw new NotImplementedException();
             }
         }
+
+        public static bool ValidateSignedData(Address addr, string signedData, string random, string data)
+        {
+            var msgData = Base16.Decode(data);
+            var randomBytes = Base16.Decode(random);
+            var signedDataBytes = Base16.Decode(signedData);
+            var msgBytes = ByteArrayUtils.ConcatBytes(randomBytes, msgData);
+            using (var stream = new MemoryStream(signedDataBytes))
+            using (var reader = new BinaryReader(stream))
+            {
+                var signature = reader.ReadSignature();
+                return signature.Verify(msgBytes, addr);
+            }
+        }
     }
 }
+
